Created folders for UML diagrams and sub-folder with User Stories uploaded
I've added an extra folder to be used for storing and organising any UML diagrams we create for the project. By storing them here, not only do we keep them where we all know they'll be if needed, but anyone external of the team can view them to see what steps we are taking.