Organisation of files, removal of unused solutions, MVC implementation, refactoring
Primarily an organisation of the file structure and changing of files inside. Removed two solutions which were not in use and then implemented the Model-View-Controller structure within the correct project files. Changed the ARP form from "main" to it's own, with a main control panel to replace it as the entry point for the program.

Refactored some files/code with new names to suit the MVC changes.