Tidied up UI, added dashboard (future use)
Used the start of this branch to tidy up the existing UI. Made the tab controls in line rather than gaining a border around them or between them and the main area. Moved parts to the edge as well to remove a border. Added a new default tab, Dashboard, to be used as the initial tab when loading. This will display information to the users, which on start up will be the previous data (so they can see what has changed from the last use).