Added ports list file, moved data table generation into Controllers
Created a first draft of the ports file. This contains port numbers, a brief description, and a recommendation of whether to leave it open, considering closing it, or to close it.
Moved the data table generation from into their respective controllers as this is the appropriate place and also helps keep frmMain.xaml.cs cleaner and easier to read.