Ports list filereader, more meaningful data for Port Scanner
Set up the file reader for the Ports List containing a description and recommendation. Edited the Data Table creation in PortScanController to include two new headers "Description" and "Recommendation" which will store these. "Local address" has been changed to "Port" and the local address portion is removed as this isn't relevant. When creating the rows of the data table, it checks if the port number that's listed has a record in the csv file and if so, it adds the description and recommendation. This has made the Port Scan output a lot more useful and meaningful to the end users. The final step for this is to tally the number of ports per recommendation and interpret this.