Port scanner completed
Port scanner functionality has been finished. The use now generates a data table, which now removes results with no port number. These are then given a description and recommendation from a CSV file of information (unknown ports receive a default message and recommendation of could be closed). The amount of each recommendation is then displayed along with information of why they have received that recommendation.

The data to be displayed from this will be the "Should be closed" values will be added to the dashboard risks section, under "Medium Risks".