Password testing finished
Functionality finished. Users enter a password and receive an estimated time to crack. They also get ticks or crosses next to commonly used strong password criteria. There is a brief description of the functionality as well as a disclaimer. Passwords are hidden by default but can be viewed in clear text.