Username search working
Username search tab is set up and working. It constructs a url and searches for the username and if the website returns an Ok status, it determines that the username exists. All matches are added to a data table which then gets displayed in a grid in the username tab.

Also removed some commented out code in ARP controller I forgot to remove previously.