Username search functionality completed
Username search works fully. Only improvement needed is using threading, the same as Nmap, so users can do other things whilst it runs as it takes some time, longer with every website added. Current method of doing this is commented out as it runs but does nothing. Will look into it tomorrow to see why it is getting caught up in itself.

Tweaked ARPController for returning "Known" devices, as it was counting registered devices as well, making the figures incorrect.