Username search now using threading
Added threading on method call in frmMain for username searching. Took a little whilst longer as it wasn't running properly, either caused by passing the username in directly through the method or the task not running long enough (Task.Delay(5000).Wait() seems to have done the trick.

Next step is to sort out the display and include figures of how many results were found vs. the amount of sites checked.