Set up dashboard functionality
Dashboard controller now takes in the ARP and Port Scan results and updates totals on the dashboard. As well as this, it also changes the risk level information based on the individual tools reports (i.e. if port scanning has 0 risks for that level, it will say it has none but the ARP will discuss the risks).

Code has been commented through.